+++ includes/tpw-core-loader.php
@@ -18,6 +18,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 require_once TPW_CORE_PATH . 'includes/class-tpw-core-activator.php';
 require_once TPW_CORE_PATH . 'includes/class-tpw-core-deactivator.php';
 require_once TPW_CORE_PATH . 'includes/class-tpw-core.php';
+require_once TPW_CORE_PATH . 'includes/scheduler/class-tpw-core-scheduler.php';
 require_once TPW_CORE_PATH . 'includes/tpw-core-functions.php';
 require_once TPW_CORE_PATH . 'includes/class-tpw-core-create-menu.php';
 require_once TPW_CORE_PATH . 'modules/costs/class-tpw-costs-save.php';